Refactor keyboard navigation; add it to settings
