+++ apps/browser/src/popup/components/navigatable-list-item.component.html
