+++ apps/browser/src/popup/components/navigatable-list.component.html
@@ -1,0 +1,3 @@
+<ng-container>
+  <ng-content></ng-content>
+</ng-container>
