+++ apps/browser/src/vault/popup/components/vault/current-tab.component.html
@@ -57,15 +57,16 @@
           <span class="flex-right">{{ loginCiphers.length }}</span>
         </h2>
         <div class="box-content">
-          <app-cipher-row
-            *ngFor="let loginCipher of loginCiphers"
-            [cipher]="loginCipher"
-            title="{{ 'autoFill' | i18n }}"
-            [showView]="true"
-            (onSelected)="fillCipher($event)"
-            (onView)="viewCipher($event)"
-          >
-          </app-cipher-row>
+          <navigatable-list-item *ngFor="let loginCipher of loginCiphers">
+            <app-cipher-row
+              [cipher]="loginCipher"
+              title="{{ 'autoFill' | i18n }}"
+              [showView]="true"
+              (onSelected)="fillCipher($event)"
+              (onView)="viewCipher($event)"
+            >
+            </app-cipher-row>
+          </navigatable-list-item>
           <div class="box-content-row padded no-hover" *ngIf="!loginCiphers.length">
             <p class="text-center">{{ "autoFillInfo" | i18n }}</p>
             <button type="button" class="btn primary link block" (click)="addCipher()">
@@ -80,14 +81,15 @@
           <span class="flex-right">{{ cardCiphers.length }}</span>
         </h2>
         <div class="box-content">
-          <app-cipher-row
-            *ngFor="let cardCipher of cardCiphers"
-            [cipher]="cardCipher"
-            title="{{ 'autoFill' | i18n }}"
-            [showView]="true"
-            (onSelected)="fillCipher($event)"
-            (onView)="viewCipher($event)"
-          ></app-cipher-row>
+          <navigatable-list-item *ngFor="let cardCipher of cardCiphers">
+            <app-cipher-row
+              [cipher]="cardCipher"
+              title="{{ 'autoFill' | i18n }}"
+              [showView]="true"
+              (onSelected)="fillCipher($event)"
+              (onView)="viewCipher($event)"
+            ></app-cipher-row>
+          </navigatable-list-item>
         </div>
       </div>
       <div class="box list" *ngIf="identityCiphers && identityCiphers.length">
@@ -96,14 +98,15 @@
           <span class="flex-right">{{ identityCiphers.length }}</span>
         </h2>
         <div class="box-content">
-          <app-cipher-row
-            *ngFor="let identityCipher of identityCiphers"
-            [cipher]="identityCipher"
-            title="{{ 'autoFill' | i18n }}"
-            [showView]="true"
-            (onSelected)="fillCipher($event)"
-            (onView)="viewCipher($event)"
-          ></app-cipher-row>
+          <navigatable-list-item *ngFor="let identityCipher of identityCiphers">
+            <app-cipher-row
+              [cipher]="identityCipher"
+              title="{{ 'autoFill' | i18n }}"
+              [showView]="true"
+              (onSelected)="fillCipher($event)"
+              (onView)="viewCipher($event)"
+            ></app-cipher-row>
+          </navigatable-list-item>
         </div>
       </div>
     </navigatable-list>
