+++ apps/browser/src/vault/popup/components/vault/vault-items.component.html
@@ -107,15 +107,18 @@
           {{ groupingTitle }}
           <span class="flex-right">{{ isSearching() ? ciphers.length : ciphers.length }}</span>
         </h2>
-        <navigatable-list class="box-content">
-          <app-cipher-row
-            *cdkVirtualFor="let c of ciphers; let last = last"
-            [cipher]="c"
-            [last]="last"
-            title="{{ 'viewItem' | i18n }}"
-            (onSelected)="selectCipher($event)"
-            (launchEvent)="launchCipher($event)"
-          ></app-cipher-row>
+        <navigatable-list>
+          <div class="box-content">
+            <navigatable-list-item *cdkVirtualFor="let c of ciphers; let last = last">
+              <app-cipher-row
+                [cipher]="c"
+                [last]="last"
+                title="{{ 'viewItem' | i18n }}"
+                (onSelected)="selectCipher($event)"
+                (launchEvent)="launchCipher($event)"
+              ></app-cipher-row>
+            </navigatable-list-item>
+          </div>
         </navigatable-list>
       </div>
     </cdk-virtual-scroll-viewport>
